requirements.txt
bandit
fire
hydra-core
mypy
pre-commit<4
pytest
pytest-cov
pytest-mock
ruff
types-PyYAML
